@@ -1,22 +1,59 @@
-import React, { Component } from 'react'
+import React, { useState, useRef } from 'react'
 import PropTypes from 'prop-types'
-
 import styles from './styles.css'
 
-export default class ExampleComponent extends Component {
-  static propTypes = {
-    text: PropTypes.string
+const TagSlider = ({ tags, addTag }) => {
+  // const { movePos, setMovePos } = setState({})
+  const [moveFlag, setMoveFlag] = useState(false)
+  const [startPos, setstartPos] = useState(0)
+  const sliderEl = useRef(null)
+
+  const mouseDownHandle = (e) => {
+    const clientX = e.clientX || e.touches[0].clientX
+
+    setMoveFlag(true)
+    setstartPos(clientX)
   }
 
-  render() {
-    const {
-      text
-    } = this.props
+  const mouseMoveHandle = (e) => {
+    if (moveFlag) {
+      const clientX = e.clientX || e.touches[0].clientX
+      sliderEl.current.scrollTo(sliderEl.current.scrollLeft + startPos - clientX, 0)
+      setstartPos(clientX)
+    }
+  }
 
-    return (
-      <div className={styles.test}>
-        Example Component: {text}
+  const mouseUpHandle = (e) => {
+    setMoveFlag(false)
+  }
+
+  return (
+    <div className={styles.tagContainer} ref={sliderEl}>
+      <div
+        className={styles.tagItems}
+        onMouseDown={mouseDownHandle}
+        onMouseMove={mouseMoveHandle}
+        onMouseUp={mouseUpHandle}
+        onMouseLeave={mouseUpHandle}
+        onTouchStart={mouseDownHandle}
+        onTouchMove={mouseMoveHandle}
+        onTouchEnd={mouseUpHandle}
+      >
+        {
+          tags.map((tag) =>
+            <div key={`tag-slider-${tag}`} className={`${styles.tagItem} ${styles.noselect}`}>
+              <span className={styles.tagName}>{tag}</span>
+            </div>
+          )
+        }
       </div>
-    )
-  }
+    </div>
+  )
 }
+
+TagSlider.propTypes = {
+  tags: PropTypes.arrayOf(PropTypes.string).isRequired,
+  addTag: PropTypes.func
+}
+
+export default TagSlider
