@@ -2,8 +2,8 @@ import babel from 'rollup-plugin-babel'
 import commonjs from 'rollup-plugin-commonjs'
 import external from 'rollup-plugin-peer-deps-external'
 import postcss from 'rollup-plugin-postcss'
-import resolve from 'rollup-plugin-node-resolve'
-import url from 'rollup-plugin-url'
+import resolve from '@rollup/plugin-node-resolve'
+import url from '@rollup/plugin-url'
 import svgr from '@svgr/rollup'
 
 import pkg from './package.json'
@@ -29,10 +29,7 @@ export default {
     }),
     url(),
     svgr(),
-    babel({
-      exclude: 'node_modules/**',
-      plugins: [ 'external-helpers' ]
-    }),
+    babel(),
     resolve(),
     commonjs()
   ]
